separate middleware, generate token for endchess
