+++ .gitignore
@@ -10,6 +10,8 @@ __data__/
 
 mongo/
 
+keys/
+
 #Build
 build/
