--- fake_login_keys/public_key.pem
@@ -1,9 +1,0 @@
------BEGIN PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAzrWcCt6hk6bNvAC7ACLo
-eBzrdEKpciBMFRJuvXikNCjLz8kdGdiKv7zTlEUcccxO8kgHdusrWcTL1UZAom4d
-0c9al5gY76+ZChjbbSFrXlXolfJyjSXx3StoJbcLJHak+aq0l9yMBhoWmkf6j2oh
-gz/p17joGmXwP9PNqLOSQ8jbIIX+q/dLWhFyh26xUh7/3rd5uYIDNL18D+W+uPRq
-zVsCcg4JZDJ/8lzjG7nwY/Z2F7css1tx+RL8yYgrnmExb+yuS0rH+xQcSdHxSRbg
-bDzkuPZZdczd+Qxauj3XGpYSYB0ZHA6VQV/kMdd87AzILC0hVltlG5bEheQ717ME
-ewIDAQAB
------END PUBLIC KEY-----
